V0.3.8(관리자페이지 공지사항 구현 중)

+++ healthme/backend/src/main/java/com/example/healthme/domain/admin/controller/ProductController.java
@@ -23,7 +23,7 @@ public class ProductController {
     }
 
     @PutMapping("/{no}")
-    public ResponseEntity<?> update_product(@PathVariable int no, @ModelAttribute InsertProductDto insertProductDto){
+    public ResponseEntity<?> update_product(@PathVariable Long no, @ModelAttribute InsertProductDto insertProductDto){
         System.out.println(insertProductDto);
         System.out.println(no);
         if(productService.update_product(no,insertProductDto)){
@@ -34,7 +34,7 @@ public class ProductController {
 
 
     @PostMapping("/delete")
-    public ResponseEntity<?> product_Delete(@RequestBody int noArray[]){
+    public ResponseEntity<?> product_Delete(@RequestBody Long noArray[]){
         boolean result = productService.product_delete(noArray);
         if (result){
             return ResponseEntity.ok("삭제완료");
